Refactor module_utils/crypto.py: split up into multiple smaller modules
