+++ plugins/modules/openssh_cert.py
@@ -208,7 +208,7 @@ from shutil import copy2, rmtree
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils._text import to_native
 
-from ansible_collections.community.crypto.plugins.module_utils.crypto import convert_relative_to_datetime
+from ansible_collections.community.crypto.plugins.module_utils.crypto.support import convert_relative_to_datetime
 
 
 class CertificateError(Exception):
