+++ plugins/modules/x509_certificate.py
@@ -861,9 +861,33 @@ from random import randint
 from ansible.module_utils.basic import AnsibleModule, missing_required_lib
 from ansible.module_utils._text import to_native, to_bytes, to_text
 
-from ansible_collections.community.crypto.plugins.module_utils import crypto as crypto_utils
 from ansible_collections.community.crypto.plugins.module_utils.compat import ipaddress as compat_ipaddress
 from ansible_collections.community.crypto.plugins.module_utils.ecs.api import ECSClient, RestOperationException, SessionConfigurationException
+
+from ansible_collections.community.crypto.plugins.module_utils.crypto.basic import (
+    OpenSSLObjectError,
+    OpenSSLBadPassphraseError,
+    write_file,
+    load_file_if_exists,
+)
+
+from ansible_collections.community.crypto.plugins.module_utils.crypto.support import (
+    OpenSSLObject,
+    load_privatekey,
+    load_certificate,
+    load_certificate_request,
+    parse_name_field,
+    get_relative_time_option,
+    select_message_digest,
+)
+
+from ansible_collections.community.crypto.plugins.module_utils.crypto.cryptography_support import (
+    cryptography_compare_public_keys,
+    cryptography_get_name,
+    cryptography_name_to_oid,
+    cryptography_key_needs_digest_for_signing,
+    cryptography_parse_key_usage_params,
+)
 
 MINIMAL_CRYPTOGRAPHY_VERSION = '1.6'
 MINIMAL_PYOPENSSL_VERSION = '0.15'
@@ -895,11 +919,11 @@ else:
     CRYPTOGRAPHY_FOUND = True
 
 
-class CertificateError(crypto_utils.OpenSSLObjectError):
+class CertificateError(OpenSSLObjectError):
     pass
 
 
-class Certificate(crypto_utils.OpenSSLObject):
+class Certificate(OpenSSLObject):
 
     def __init__(self, module, backend):
         super(Certificate, self).__init__(
@@ -945,7 +969,7 @@ class Certificate(crypto_utils.OpenSSLObject):
             except OpenSSL.SSL.Error:
                 return False
         elif self.backend == 'cryptography':
-            return crypto_utils.cryptography_compare_public_keys(self.cert.public_key(), self.privatekey.public_key())
+            return cryptography_compare_public_keys(self.cert.public_key(), self.privatekey.public_key())
 
     def _validate_csr(self):
         if self.backend == 'pyopenssl':
@@ -972,7 +996,7 @@ class Certificate(crypto_utils.OpenSSLObject):
             # Verify that CSR is signed by certificate's private key
             if not self.csr.is_signature_valid:
                 return False
-            if not crypto_utils.cryptography_compare_public_keys(self.csr.public_key(), self.cert.public_key()):
+            if not cryptography_compare_public_keys(self.csr.public_key(), self.cert.public_key()):
                 return False
             # Check subject
             if self.csr.subject != self.cert.subject:
@@ -1013,25 +1037,25 @@ class Certificate(crypto_utils.OpenSSLObject):
             return False
 
         try:
-            self.cert = crypto_utils.load_certificate(self.path, backend=self.backend)
+            self.cert = load_certificate(self.path, backend=self.backend)
         except Exception as dummy:
             return False
 
         if self.privatekey_path or self.privatekey_content:
             try:
-                self.privatekey = crypto_utils.load_privatekey(
+                self.privatekey = load_privatekey(
                     path=self.privatekey_path,
                     content=self.privatekey_content,
                     passphrase=self.privatekey_passphrase,
                     backend=self.backend
                 )
-            except crypto_utils.OpenSSLBadPassphraseError as exc:
+            except OpenSSLBadPassphraseError as exc:
                 raise CertificateError(exc)
             if not self._validate_privatekey():
                 return False
 
         if self.csr_path or self.csr_content:
-            self.csr = crypto_utils.load_certificate_request(
+            self.csr = load_certificate_request(
                 path=self.csr_path,
                 content=self.csr_content,
                 backend=self.backend
@@ -1094,9 +1118,9 @@ class SelfSignedCertificateCryptography(Certificate):
     def __init__(self, module):
         super(SelfSignedCertificateCryptography, self).__init__(module, 'cryptography')
         self.create_subject_key_identifier = module.params['selfsigned_create_subject_key_identifier']
-        self.notBefore = crypto_utils.get_relative_time_option(module.params['selfsigned_not_before'], 'selfsigned_not_before', backend=self.backend)
-        self.notAfter = crypto_utils.get_relative_time_option(module.params['selfsigned_not_after'], 'selfsigned_not_after', backend=self.backend)
-        self.digest = crypto_utils.select_message_digest(module.params['selfsigned_digest'])
+        self.notBefore = get_relative_time_option(module.params['selfsigned_not_before'], 'selfsigned_not_before', backend=self.backend)
+        self.notAfter = get_relative_time_option(module.params['selfsigned_not_after'], 'selfsigned_not_after', backend=self.backend)
+        self.digest = select_message_digest(module.params['selfsigned_digest'])
         self.version = module.params['selfsigned_version']
         self.serial_number = x509.random_serial_number()
 
@@ -1109,7 +1133,7 @@ class SelfSignedCertificateCryptography(Certificate):
                 'The private key file {0} does not exist'.format(self.privatekey_path)
             )
 
-        self.csr = crypto_utils.load_certificate_request(
+        self.csr = load_certificate_request(
             path=self.csr_path,
             content=self.csr_content,
             backend=self.backend
@@ -1117,16 +1141,16 @@ class SelfSignedCertificateCryptography(Certificate):
         self._module = module
 
         try:
-            self.privatekey = crypto_utils.load_privatekey(
+            self.privatekey = load_privatekey(
                 path=self.privatekey_path,
                 content=self.privatekey_content,
                 passphrase=self.privatekey_passphrase,
                 backend=self.backend
             )
-        except crypto_utils.OpenSSLBadPassphraseError as exc:
+        except OpenSSLBadPassphraseError as exc:
             module.fail_json(msg=to_native(exc))
 
-        if crypto_utils.cryptography_key_needs_digest_for_signing(self.privatekey):
+        if cryptography_key_needs_digest_for_signing(self.privatekey):
             if self.digest is None:
                 raise CertificateError(
                     'The digest %s is not supported with the cryptography backend' % module.params['selfsigned_digest']
@@ -1181,10 +1205,10 @@ class SelfSignedCertificateCryptography(Certificate):
 
             if self.backup:
                 self.backup_file = module.backup_local(self.path)
-            crypto_utils.write_file(module, certificate.public_bytes(Encoding.PEM))
+            write_file(module, certificate.public_bytes(Encoding.PEM))
             self.changed = True
         else:
-            self.cert = crypto_utils.load_certificate(self.path, backend=self.backend)
+            self.cert = load_certificate(self.path, backend=self.backend)
 
         file_args = module.load_file_common_arguments(module.params)
         if module.set_fs_attributes_if_different(file_args, False):
@@ -1201,7 +1225,7 @@ class SelfSignedCertificateCryptography(Certificate):
         if self.backup_file:
             result['backup_file'] = self.backup_file
         if self.return_content:
-            content = crypto_utils.load_file_if_exists(self.path, ignore_errors=True)
+            content = load_file_if_exists(self.path, ignore_errors=True)
             result['certificate'] = content.decode('utf-8') if content else None
 
         if check_mode:
@@ -1227,8 +1251,8 @@ class SelfSignedCertificate(Certificate):
         super(SelfSignedCertificate, self).__init__(module, 'pyopenssl')
         if module.params['selfsigned_create_subject_key_identifier'] != 'create_if_not_provided':
             module.fail_json(msg='selfsigned_create_subject_key_identifier cannot be used with the pyOpenSSL backend!')
-        self.notBefore = crypto_utils.get_relative_time_option(module.params['selfsigned_not_before'], 'selfsigned_not_before', backend=self.backend)
-        self.notAfter = crypto_utils.get_relative_time_option(module.params['selfsigned_not_after'], 'selfsigned_not_after', backend=self.backend)
+        self.notBefore = get_relative_time_option(module.params['selfsigned_not_before'], 'selfsigned_not_before', backend=self.backend)
+        self.notAfter = get_relative_time_option(module.params['selfsigned_not_after'], 'selfsigned_not_after', backend=self.backend)
         self.digest = module.params['selfsigned_digest']
         self.version = module.params['selfsigned_version']
         self.serial_number = randint(1000, 99999)
@@ -1242,17 +1266,17 @@ class SelfSignedCertificate(Certificate):
                 'The private key file {0} does not exist'.format(self.privatekey_path)
             )
 
-        self.csr = crypto_utils.load_certificate_request(
+        self.csr = load_certificate_request(
             path=self.csr_path,
             content=self.csr_content,
         )
         try:
-            self.privatekey = crypto_utils.load_privatekey(
+            self.privatekey = load_privatekey(
                 path=self.privatekey_path,
                 content=self.privatekey_content,
                 passphrase=self.privatekey_passphrase,
             )
-        except crypto_utils.OpenSSLBadPassphraseError as exc:
+        except OpenSSLBadPassphraseError as exc:
             module.fail_json(msg=str(exc))
 
     def generate(self, module):
@@ -1283,7 +1307,7 @@ class SelfSignedCertificate(Certificate):
 
             if self.backup:
                 self.backup_file = module.backup_local(self.path)
-            crypto_utils.write_file(module, crypto.dump_certificate(crypto.FILETYPE_PEM, self.cert))
+            write_file(module, crypto.dump_certificate(crypto.FILETYPE_PEM, self.cert))
             self.changed = True
 
         file_args = module.load_file_common_arguments(module.params)
@@ -1301,7 +1325,7 @@ class SelfSignedCertificate(Certificate):
         if self.backup_file:
             result['backup_file'] = self.backup_file
         if self.return_content:
-            content = crypto_utils.load_file_if_exists(self.path, ignore_errors=True)
+            content = load_file_if_exists(self.path, ignore_errors=True)
             result['certificate'] = content.decode('utf-8') if content else None
 
         if check_mode:
@@ -1326,9 +1350,9 @@ class OwnCACertificateCryptography(Certificate):
         super(OwnCACertificateCryptography, self).__init__(module, 'cryptography')
         self.create_subject_key_identifier = module.params['ownca_create_subject_key_identifier']
         self.create_authority_key_identifier = module.params['ownca_create_authority_key_identifier']
-        self.notBefore = crypto_utils.get_relative_time_option(module.params['ownca_not_before'], 'ownca_not_before', backend=self.backend)
-        self.notAfter = crypto_utils.get_relative_time_option(module.params['ownca_not_after'], 'ownca_not_after', backend=self.backend)
-        self.digest = crypto_utils.select_message_digest(module.params['ownca_digest'])
+        self.notBefore = get_relative_time_option(module.params['ownca_not_before'], 'ownca_not_before', backend=self.backend)
+        self.notAfter = get_relative_time_option(module.params['ownca_not_after'], 'ownca_not_after', backend=self.backend)
+        self.digest = select_message_digest(module.params['ownca_digest'])
         self.version = module.params['ownca_version']
         self.serial_number = x509.random_serial_number()
         self.ca_cert_path = module.params['ownca_path']
@@ -1354,27 +1378,27 @@ class OwnCACertificateCryptography(Certificate):
                 'The CA private key file {0} does not exist'.format(self.ca_privatekey_path)
             )
 
-        self.csr = crypto_utils.load_certificate_request(
+        self.csr = load_certificate_request(
             path=self.csr_path,
             content=self.csr_content,
             backend=self.backend
         )
-        self.ca_cert = crypto_utils.load_certificate(
+        self.ca_cert = load_certificate(
             path=self.ca_cert_path,
             content=self.ca_cert_content,
             backend=self.backend
         )
         try:
-            self.ca_private_key = crypto_utils.load_privatekey(
+            self.ca_private_key = load_privatekey(
                 path=self.ca_privatekey_path,
                 content=self.ca_privatekey_content,
                 passphrase=self.ca_privatekey_passphrase,
                 backend=self.backend
             )
-        except crypto_utils.OpenSSLBadPassphraseError as exc:
+        except OpenSSLBadPassphraseError as exc:
             module.fail_json(msg=str(exc))
 
-        if crypto_utils.cryptography_key_needs_digest_for_signing(self.ca_private_key):
+        if cryptography_key_needs_digest_for_signing(self.ca_private_key):
             if self.digest is None:
                 raise CertificateError(
                     'The digest %s is not supported with the cryptography backend' % module.params['ownca_digest']
@@ -1450,10 +1474,10 @@ class OwnCACertificateCryptography(Certificate):
 
             if self.backup:
                 self.backup_file = module.backup_local(self.path)
-            crypto_utils.write_file(module, certificate.public_bytes(Encoding.PEM))
+            write_file(module, certificate.public_bytes(Encoding.PEM))
             self.changed = True
         else:
-            self.cert = crypto_utils.load_certificate(self.path, backend=self.backend)
+            self.cert = load_certificate(self.path, backend=self.backend)
 
         file_args = module.load_file_common_arguments(module.params)
         if module.set_fs_attributes_if_different(file_args, False):
@@ -1498,7 +1522,7 @@ class OwnCACertificateCryptography(Certificate):
         if self.backup_file:
             result['backup_file'] = self.backup_file
         if self.return_content:
-            content = crypto_utils.load_file_if_exists(self.path, ignore_errors=True)
+            content = load_file_if_exists(self.path, ignore_errors=True)
             result['certificate'] = content.decode('utf-8') if content else None
 
         if check_mode:
@@ -1522,8 +1546,8 @@ class OwnCACertificate(Certificate):
 
     def __init__(self, module):
         super(OwnCACertificate, self).__init__(module, 'pyopenssl')
-        self.notBefore = crypto_utils.get_relative_time_option(module.params['ownca_not_before'], 'ownca_not_before', backend=self.backend)
-        self.notAfter = crypto_utils.get_relative_time_option(module.params['ownca_not_after'], 'ownca_not_after', backend=self.backend)
+        self.notBefore = get_relative_time_option(module.params['ownca_not_before'], 'ownca_not_before', backend=self.backend)
+        self.notAfter = get_relative_time_option(module.params['ownca_not_after'], 'ownca_not_after', backend=self.backend)
         self.digest = module.params['ownca_digest']
         self.version = module.params['ownca_version']
         self.serial_number = randint(1000, 99999)
@@ -1554,21 +1578,21 @@ class OwnCACertificate(Certificate):
                 'The CA private key file {0} does not exist'.format(self.ca_privatekey_path)
             )
 
-        self.csr = crypto_utils.load_certificate_request(
+        self.csr = load_certificate_request(
             path=self.csr_path,
             content=self.csr_content,
         )
-        self.ca_cert = crypto_utils.load_certificate(
+        self.ca_cert = load_certificate(
             path=self.ca_cert_path,
             content=self.ca_cert_content,
         )
         try:
-            self.ca_privatekey = crypto_utils.load_privatekey(
+            self.ca_privatekey = load_privatekey(
                 path=self.ca_privatekey_path,
                 content=self.ca_privatekey_content,
                 passphrase=self.ca_privatekey_passphrase
             )
-        except crypto_utils.OpenSSLBadPassphraseError as exc:
+        except OpenSSLBadPassphraseError as exc:
             module.fail_json(msg=str(exc))
 
     def generate(self, module):
@@ -1604,7 +1628,7 @@ class OwnCACertificate(Certificate):
 
             if self.backup:
                 self.backup_file = module.backup_local(self.path)
-            crypto_utils.write_file(module, crypto.dump_certificate(crypto.FILETYPE_PEM, self.cert))
+            write_file(module, crypto.dump_certificate(crypto.FILETYPE_PEM, self.cert))
             self.changed = True
 
         file_args = module.load_file_common_arguments(module.params)
@@ -1624,7 +1648,7 @@ class OwnCACertificate(Certificate):
         if self.backup_file:
             result['backup_file'] = self.backup_file
         if self.return_content:
-            content = crypto_utils.load_file_if_exists(self.path, ignore_errors=True)
+            content = load_file_if_exists(self.path, ignore_errors=True)
             result['certificate'] = content.decode('utf-8') if content else None
 
         if check_mode:
@@ -1667,12 +1691,12 @@ class AssertOnlyCertificateBase(Certificate):
 
         self.signature_algorithms = module.params['signature_algorithms']
         if module.params['subject']:
-            self.subject = crypto_utils.parse_name_field(module.params['subject'])
+            self.subject = parse_name_field(module.params['subject'])
         else:
             self.subject = []
         self.subject_strict = module.params['subject_strict']
         if module.params['issuer']:
-            self.issuer = crypto_utils.parse_name_field(module.params['issuer'])
+            self.issuer = parse_name_field(module.params['issuer'])
         else:
             self.issuer = []
         self.issuer_strict = module.params['issuer_strict']
@@ -1697,19 +1721,19 @@ class AssertOnlyCertificateBase(Certificate):
             self.valid_in = "+" + self.valid_in + "s"
 
         # Load objects
-        self.cert = crypto_utils.load_certificate(self.path, backend=self.backend)
+        self.cert = load_certificate(self.path, backend=self.backend)
         if self.privatekey_path is not None or self.privatekey_content is not None:
             try:
-                self.privatekey = crypto_utils.load_privatekey(
+                self.privatekey = load_privatekey(
                     path=self.privatekey_path,
                     content=self.privatekey_content,
                     passphrase=self.privatekey_passphrase,
                     backend=self.backend
                 )
-            except crypto_utils.OpenSSLBadPassphraseError as exc:
+            except OpenSSLBadPassphraseError as exc:
                 raise CertificateError(exc)
         if self.csr_path is not None or self.csr_content is not None:
-            self.csr = crypto_utils.load_certificate_request(
+            self.csr = load_certificate_request(
                 path=self.csr_path,
                 content=self.csr_content,
                 backend=self.backend
@@ -1884,7 +1908,7 @@ class AssertOnlyCertificateBase(Certificate):
 
         if self.not_before is not None:
             cert_not_valid_before = self._validate_not_before()
-            if cert_not_valid_before != crypto_utils.get_relative_time_option(self.not_before, 'not_before', backend=self.backend):
+            if cert_not_valid_before != get_relative_time_option(self.not_before, 'not_before', backend=self.backend):
                 messages.append(
                     'Invalid not_before component (got %s, expected %s to be present)' %
                     (cert_not_valid_before, self.not_before)
@@ -1892,7 +1916,7 @@ class AssertOnlyCertificateBase(Certificate):
 
         if self.not_after is not None:
             cert_not_valid_after = self._validate_not_after()
-            if cert_not_valid_after != crypto_utils.get_relative_time_option(self.not_after, 'not_after', backend=self.backend):
+            if cert_not_valid_after != get_relative_time_option(self.not_after, 'not_after', backend=self.backend):
                 messages.append(
                     'Invalid not_after component (got %s, expected %s to be present)' %
                     (cert_not_valid_after, self.not_after)
@@ -1942,7 +1966,7 @@ class AssertOnlyCertificateBase(Certificate):
             'csr': self.csr_path,
         }
         if self.return_content:
-            content = crypto_utils.load_file_if_exists(self.path, ignore_errors=True)
+            content = load_file_if_exists(self.path, ignore_errors=True)
             result['certificate'] = content.decode('utf-8') if content else None
         return result
 
@@ -1953,12 +1977,12 @@ class AssertOnlyCertificateCryptography(AssertOnlyCertificateBase):
         super(AssertOnlyCertificateCryptography, self).__init__(module, 'cryptography')
 
     def _validate_privatekey(self):
-        return crypto_utils.cryptography_compare_public_keys(self.cert.public_key(), self.privatekey.public_key())
+        return cryptography_compare_public_keys(self.cert.public_key(), self.privatekey.public_key())
 
     def _validate_csr_signature(self):
         if not self.csr.is_signature_valid:
             return False
-        return crypto_utils.cryptography_compare_public_keys(self.csr.public_key(), self.cert.public_key())
+        return cryptography_compare_public_keys(self.csr.public_key(), self.cert.public_key())
 
     def _validate_csr_subject(self):
         return self.csr.subject == self.cert.subject
@@ -1982,14 +2006,14 @@ class AssertOnlyCertificateCryptography(AssertOnlyCertificateBase):
             return self.cert.signature_algorithm_oid._name
 
     def _validate_subject(self):
-        expected_subject = Name([NameAttribute(oid=crypto_utils.cryptography_name_to_oid(sub[0]), value=to_text(sub[1]))
+        expected_subject = Name([NameAttribute(oid=cryptography_name_to_oid(sub[0]), value=to_text(sub[1]))
                                  for sub in self.subject])
         cert_subject = self.cert.subject
         if not compare_sets(expected_subject, cert_subject, self.subject_strict):
             return expected_subject, cert_subject
 
     def _validate_issuer(self):
-        expected_issuer = Name([NameAttribute(oid=crypto_utils.cryptography_name_to_oid(iss[0]), value=to_text(iss[1]))
+        expected_issuer = Name([NameAttribute(oid=cryptography_name_to_oid(iss[0]), value=to_text(iss[1]))
                                 for iss in self.issuer])
         cert_issuer = self.cert.issuer
         if not compare_sets(expected_issuer, cert_issuer, self.issuer_strict):
@@ -2027,7 +2051,7 @@ class AssertOnlyCertificateCryptography(AssertOnlyCertificateBase):
                     decipher_only=current_key_usage.decipher_only
                 ))
 
-            key_usages = crypto_utils.cryptography_parse_key_usage_params(self.key_usage)
+            key_usages = cryptography_parse_key_usage_params(self.key_usage)
             if not compare_dicts(key_usages, test_key_usage, self.key_usage_strict):
                 return self.key_usage, [k for k, v in test_key_usage.items() if v is True]
 
@@ -2039,7 +2063,7 @@ class AssertOnlyCertificateCryptography(AssertOnlyCertificateBase):
     def _validate_extended_key_usage(self):
         try:
             current_ext_keyusage = self.cert.extensions.get_extension_for_class(x509.ExtendedKeyUsage).value
-            usages = [crypto_utils.cryptography_name_to_oid(usage) for usage in self.extended_key_usage]
+            usages = [cryptography_name_to_oid(usage) for usage in self.extended_key_usage]
             expected_ext_keyusage = x509.ExtendedKeyUsage(usages)
             if not compare_sets(expected_ext_keyusage, current_ext_keyusage, self.extended_key_usage_strict):
                 return [eku.value for eku in expected_ext_keyusage], [eku.value for eku in current_ext_keyusage]
@@ -2052,7 +2076,7 @@ class AssertOnlyCertificateCryptography(AssertOnlyCertificateBase):
     def _validate_subject_alt_name(self):
         try:
             current_san = self.cert.extensions.get_extension_for_class(x509.SubjectAlternativeName).value
-            expected_san = [crypto_utils.cryptography_get_name(san) for san in self.subject_alt_name]
+            expected_san = [cryptography_get_name(san) for san in self.subject_alt_name]
             if not compare_sets(expected_san, current_san, self.subject_alt_name_strict):
                 return self.subject_alt_name, current_san
         except cryptography.x509.ExtensionNotFound:
@@ -2067,15 +2091,15 @@ class AssertOnlyCertificateCryptography(AssertOnlyCertificateBase):
         return self.cert.not_valid_after
 
     def _validate_valid_at(self):
-        rt = crypto_utils.get_relative_time_option(self.valid_at, 'valid_at', backend=self.backend)
+        rt = get_relative_time_option(self.valid_at, 'valid_at', backend=self.backend)
         return self.cert.not_valid_before, rt, self.cert.not_valid_after
 
     def _validate_invalid_at(self):
-        rt = crypto_utils.get_relative_time_option(self.invalid_at, 'invalid_at', backend=self.backend)
+        rt = get_relative_time_option(self.invalid_at, 'invalid_at', backend=self.backend)
         return self.cert.not_valid_before, rt, self.cert.not_valid_after
 
     def _validate_valid_in(self):
-        valid_in_date = crypto_utils.get_relative_time_option(self.valid_in, "valid_in", backend=self.backend)
+        valid_in_date = get_relative_time_option(self.valid_in, "valid_in", backend=self.backend)
         return self.cert.not_valid_before, valid_in_date, self.cert.not_valid_after
 
 
@@ -2232,17 +2256,17 @@ class AssertOnlyCertificate(AssertOnlyCertificateBase):
         return self.cert.get_notAfter()
 
     def _validate_valid_at(self):
-        rt = crypto_utils.get_relative_time_option(self.valid_at, "valid_at", backend=self.backend)
+        rt = get_relative_time_option(self.valid_at, "valid_at", backend=self.backend)
         rt = to_bytes(rt, errors='surrogate_or_strict')
         return self.cert.get_notBefore(), rt, self.cert.get_notAfter()
 
     def _validate_invalid_at(self):
-        rt = crypto_utils.get_relative_time_option(self.invalid_at, "invalid_at", backend=self.backend)
+        rt = get_relative_time_option(self.invalid_at, "invalid_at", backend=self.backend)
         rt = to_bytes(rt, errors='surrogate_or_strict')
         return self.cert.get_notBefore(), rt, self.cert.get_notAfter()
 
     def _validate_valid_in(self):
-        valid_in_asn1 = crypto_utils.get_relative_time_option(self.valid_in, "valid_in", backend=self.backend)
+        valid_in_asn1 = get_relative_time_option(self.valid_in, "valid_in", backend=self.backend)
         valid_in_date = to_bytes(valid_in_asn1, errors='surrogate_or_strict')
         return self.cert.get_notBefore(), valid_in_date, self.cert.get_notAfter()
 
@@ -2253,14 +2277,14 @@ class EntrustCertificate(Certificate):
     def __init__(self, module, backend):
         super(EntrustCertificate, self).__init__(module, backend)
         self.trackingId = None
-        self.notAfter = crypto_utils.get_relative_time_option(module.params['entrust_not_after'], 'entrust_not_after', backend=self.backend)
+        self.notAfter = get_relative_time_option(module.params['entrust_not_after'], 'entrust_not_after', backend=self.backend)
 
         if self.csr_content is None or not os.path.exists(self.csr_path):
             raise CertificateError(
                 'The certificate signing request file {0} does not exist'.format(self.csr_path)
             )
 
-        self.csr = crypto_utils.load_certificate_request(
+        self.csr = load_certificate_request(
             path=self.csr_path,
             content=self.csr_content,
             backend=self.backend,
@@ -2340,8 +2364,8 @@ class EntrustCertificate(Certificate):
 
             if self.backup:
                 self.backup_file = module.backup_local(self.path)
-            crypto_utils.write_file(module, to_bytes(result.get('endEntityCert')))
-            self.cert = crypto_utils.load_certificate(self.path, backend=self.backend)
+            write_file(module, to_bytes(result.get('endEntityCert')))
+            self.cert = load_certificate(self.path, backend=self.backend)
             self.changed = True
 
     def check(self, module, perms_required=True):
@@ -2410,7 +2434,7 @@ class EntrustCertificate(Certificate):
         if self.backup_file:
             result['backup_file'] = self.backup_file
         if self.return_content:
-            content = crypto_utils.load_file_if_exists(self.path, ignore_errors=True)
+            content = load_file_if_exists(self.path, ignore_errors=True)
             result['certificate'] = content.decode('utf-8') if content else None
 
         result.update(self._get_cert_details())
@@ -2481,7 +2505,7 @@ class AcmeCertificate(Certificate):
                 crt = module.run_command(command, check_rc=True)[1]
                 if self.backup:
                     self.backup_file = module.backup_local(self.path)
-                crypto_utils.write_file(module, to_bytes(crt))
+                write_file(module, to_bytes(crt))
                 self.changed = True
             except OSError as exc:
                 raise CertificateError(exc)
@@ -2502,7 +2526,7 @@ class AcmeCertificate(Certificate):
         if self.backup_file:
             result['backup_file'] = self.backup_file
         if self.return_content:
-            content = crypto_utils.load_file_if_exists(self.path, ignore_errors=True)
+            content = load_file_if_exists(self.path, ignore_errors=True)
             result['certificate'] = content.decode('utf-8') if content else None
 
         return result
@@ -2725,7 +2749,7 @@ def main():
 
         result = certificate.dump()
         module.exit_json(**result)
-    except crypto_utils.OpenSSLObjectError as exc:
+    except OpenSSLObjectError as exc:
         module.fail_json(msg=to_native(exc))
